flask-angular-mysql/flask/user-app/requirements.txt
Flask
mysql-connector
mysql-connector-python
Flask_Cors